# compilerGetMatches: improve performance