@Minoru 2021-09-10
I can't confirm:

```
$ hyperfine \
--prepare './before clean' './before build' \
--prepare './after clean' './after build'
Benchmark #1: ./before build
  Time (mean ± σ):     11.872 s ±  0.205 s    [User: 43.154 s, System: 11.354 s]
  Range (min … max):   11.482 s … 12.155 s    10 runs
 
Benchmark #2: ./after build
  Time (mean ± σ):     11.819 s ±  0.204 s    [User: 43.706 s, System: 10.957 s]
  Range (min … max):   11.503 s … 12.117 s    10 runs
 
Summary
  './after build' ran
    1.00 ± 0.02 times faster than './before build'
```

But hey, specialized functions are always better, so I'll merge this anyway :) Thanks for the PR!